third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/responsexml-type-expected.txt

This tests that XMLHttpRequest.responseXML returns an XMLDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS responseDocument is "[object XMLDocument]"
PASS successfullyParsed is true

TEST COMPLETE
